Criação da pasta "controllers" para conter os controllers referentes as tasks e aos usuarios, como também foram criados funções de delete, inserção e get dentro do controller "UserController" e está tudo funcionando perfeitamente.

+++ src/main/java/cursospringboot/example/cursospringbootjava/repositories/UserReposetery.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserReposetery extends JpaRepository<User, Long> {
+public interface UserReposetery extends JpaRepository<User, Integer> {
 
 
 }
